feat: :white_check_mark: updates and new content

+++ Dia1/Dia2/Ejercicio5_Dia2_SantosDaniel.py
@@ -1,6 +1,10 @@
+# ##########################
+# #### Clase Dia 2 ######
+# ##########################
+
 #Convertir de grados Celsius a Farenheit
 #Solicitamos que ingrese una temperatura en Celsius
-gCelsius = int(input("Ingrese una temperatura en grados Celsius: "))
+gCelsius = float(input("Ingrese una temperatura en grados Celsius: "))
 
 #Realizamos la conversion usando una formula
 gFarenheit = (gCelsius*9/5)+32
